pyproject.toml
[project]
name = "intech-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.2",
]
